% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zs.R
\name{clean_zs}
\alias{clean_zs}
\title{Clean z-score}
\usage{
clean_zs(data)
}
\arguments{
\item{data}{The z score matrix that need to be cleaned}
}
\value{
Cleaned z score matrix
}
\description{
Clean the z-score matrix through replacing the infinite values
}
