% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_zscore_estimation.R
\name{fast_zscore_estimation}
\alias{fast_zscore_estimation}
\title{fast z score estimation}
\usage{
fast_zscore_estimation(data_file)
}
\arguments{
\item{data_file}{the path of the .csv or .tsv file}
}
\value{
absolute path of the saved file storing the z_scores data frame
}
\description{
A function to estimate and generate a z-score matrix for the read count
}
