% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_size_factor.R
\name{get_size_factor}
\alias{get_size_factor}
\title{Get size factor}
\usage{
get_size_factor(df)
}
\arguments{
\item{df}{a dataframe with the gene counts.}
}
\value{
the size factor fore each column
}
\description{
Calculation of the size factor for each column (sample) as the median value
of the non_zero normalized counts. The original read counts for each are normalized
to the geometric means of the counts.
}
