% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_z_scores.R
\name{get_z_scores}
\alias{get_z_scores}
\title{get z score}
\usage{
get_z_scores(data, l_ji_other__ = NULL)
}
\arguments{
\item{data}{A size J*N dataframe}

\item{l_ji_other__}{A size *N dataframe}
}
\value{
A list containing Z scores and l_ji__
}
\description{
get z score
}
\examples{
data = read.csv("geneCounts.tsv",sep="\t")
get_z_scores(data)
}
