% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_transform.R
\name{h_transform}
\alias{h_transform}
\title{Transform data}
\usage{
h_transform(a, transform_f, axis = NULL, print_ = FALSE)
}
\arguments{
\item{a}{A matrix needed to be operated}

\item{transform_f}{The function need to be applied}

\item{axis}{The axis transform_f needed to be applied. Default: NULL}

\item{print_}{Whether to print axis message, default FALSE}
}
\value{
the size factor fore each column
}
\description{
Perform a user_identified transformation to the matrix.
}
