% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optht.R
\name{optht}
\alias{optht}
\title{OPTHT}
\usage{
optht(data, sv, sigma = NA, debug)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{sv}{A vector of singular values from the matrix}

\item{sigma}{True variance if known, is NA by default}

\item{debug}{A boolean indicating if want verbose information}
}
\value{
An index for the hardthreshold for the singular values
}
\description{
A fast implementation of simple linear regression
}
