% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optht_svd_zs.R
\name{optht_svd_zs}
\alias{optht_svd_zs}
\title{OPTHT-SVD z score estimation}
\usage{
optht_svd_zs(data_file)
}
\arguments{
\item{data_file}{the name of z_score for the read count}
}
\value{
absolute path of the saved file
}
\description{
A function to estimate and generate a z-score matrix after removing the noise using SVD and OPT, working as the outlier score.
p value will be calcualed according to the z-score and saved.
}
