% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_svd.R
\name{reverse_svd}
\alias{reverse_svd}
\title{Reverse SVD}
\usage{
reverse_svd(data, U, s, VT, s_dimension)
}
\arguments{
\item{data}{The original gene count matrix}

\item{U}{the U matrix after SVD decomposition of the original read count}

\item{s}{the singular matrix after SVD decomposition of the original read count}

\item{VT}{the VT matrix after SVD decomposition of the original read count}

\item{s_dimension}{cut off value calculated by optht function}
}
\value{
the reconstructed gene count matrix after removing noises.
}
\description{
Calculate the read count matrix after removing the noises.
}
