% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize}
\usage{
standardize(data, axis = NA)
}
\arguments{
\item{data}{A data matrix need to be standardize}

\item{axis}{The axis needed to be standardized, default NA}
}
\value{
standardized data
}
\description{
Standardize a data matrix according to the given axis into mean of 0 and a standard deviation of 1.
}
