% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{std}
\usage{
std(data, axis = NA)
}
\arguments{
\item{data}{a data matrix.}

\item{axis}{The axis to perform the calculation.}
}
\value{
Standardized standard deviation
}
\description{
Calculation the standardized standard deviation
}
